// src/components/salla-tabs/interfaces.ts
export interface ISallaTabData {
    selected: () => void;
    unselect: () => void;
    name: string;
}

export interface ISallaTabHeaderData extends ISallaTabData {
    id: string;
}

export interface ISallaTabContentData extends ISallaTabData { }


export interface ISallaTabGroup {
    header: ISallaTabHeaderData;
    content: ISallaTabContentData;
}